+++ src/api.jl
@@ -1,3 +1,15 @@
+function tcl_eval end
+function tcl_exec end
+function tcl_list end
+function tcl_concat end
+function tcl_getvar end
+function tcl_setvar end
+function tcl_unsetvar end
+function tcl_exists end
+function tcl_quote_string end
+function tcl_version end
+function tcl_library end
+
 export wm
 
 """